ENH: Restore EarlyStoppingHook state from checkpoint.
This should fix #239 by making two changes:
- Store the best loss in the hook's `state_dict`.
- Ensuring `Trainer` uses the actual hook `state_dict` for restore.

Note that the changes to `Trainer` might break some
workflows that involve adding or removing hooks before
restoring from a checkpoint: this will now fail if the
`state_dict` that is restored doesn't match the new hook
setup. There should probably be a check to see if the setups
match... or a separate method to do a "less thorough" restore.